requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
pydantic==1.10.7
pika==1.3.1
python-dotenv==1.0.0 